+++ x/btcstaking/types/msg_test.go
@@ -10,6 +10,8 @@ import (
 	"github.com/babylonlabs-io/babylon/v3/testutil/datagen"
 	bbntypes "github.com/babylonlabs-io/babylon/v3/types"
 	"github.com/babylonlabs-io/babylon/v3/x/btcstaking/types"
+	"github.com/btcsuite/btcd/btcec/v2"
+	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	stktypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/stretchr/testify/require"
@@ -185,6 +187,7 @@ func TestMsgCreateFinalityProviderValidateBasic(t *testing.T) {
 func TestMsgEditFinalityProviderValidateBasic(t *testing.T) {
 	var (
 		r                = rand.New(rand.NewSource(10))
+		addr             = datagen.GenRandomAddress().String()
 		randomDecPointer = func() *math.LegacyDec {
 			val := datagen.RandomLegacyDec(r, 10, 1)
 			return &val
@@ -205,6 +208,7 @@ func TestMsgEditFinalityProviderValidateBasic(t *testing.T) {
 		{
 			name: "valid commission and description",
 			msg: &types.MsgEditFinalityProvider{
+				Addr:        addr,
 				Commission:  randomDecPointer(),
 				Description: fpDesc,
 				BtcPk:       []byte(*validPk),
@@ -214,6 +218,7 @@ func TestMsgEditFinalityProviderValidateBasic(t *testing.T) {
 		{
 			name: "commission negative value",
 			msg: &types.MsgEditFinalityProvider{
+				Addr:        addr,
 				Commission:  &negativeDec,
 				Description: fpDesc,
 				BtcPk:       []byte(*validPk),
@@ -223,6 +228,7 @@ func TestMsgEditFinalityProviderValidateBasic(t *testing.T) {
 		{
 			name: "commission greater than 1",
 			msg: &types.MsgEditFinalityProvider{
+				Addr:        addr,
 				Commission:  &biggerThanOneDec,
 				Description: fpDesc,
 				BtcPk:       []byte(*validPk),
@@ -232,6 +238,7 @@ func TestMsgEditFinalityProviderValidateBasic(t *testing.T) {
 		{
 			name: "empty description",
 			msg: &types.MsgEditFinalityProvider{
+				Addr:        addr,
 				Description: nil,
 				BtcPk:       []byte(*validPk),
 			},
@@ -240,6 +247,7 @@ func TestMsgEditFinalityProviderValidateBasic(t *testing.T) {
 		{
 			name: "empty moniker",
 			msg: &types.MsgEditFinalityProvider{
+				Addr: addr,
 				Description: &stktypes.Description{
 					Moniker: "",
 				},
@@ -250,6 +258,7 @@ func TestMsgEditFinalityProviderValidateBasic(t *testing.T) {
 		{
 			name: "invalid BTC public key length",
 			msg: &types.MsgEditFinalityProvider{
+				Addr:        addr,
 				Description: fpDesc,
 				BtcPk:       []byte("shortBTCpk"),
 			},
@@ -258,10 +267,20 @@ func TestMsgEditFinalityProviderValidateBasic(t *testing.T) {
 		{
 			name: "invalid BTC public key (non-hex)",
 			msg: &types.MsgEditFinalityProvider{
+				Addr:        addr,
 				Description: fpDesc,
 				BtcPk:       []byte("B3C0F1D2E3A4B596C7D8E9FA1B2C3D4E5F6A7B8C9D0E1F2A3B4C5D6E7F8A9BZ"),
 			},
 			expected: fmt.Errorf("malformed BTC PK"),
+		},
+		{
+			name: "empty FP addr",
+			msg: &types.MsgEditFinalityProvider{
+				Commission:  randomDecPointer(),
+				Description: fpDesc,
+				BtcPk:       []byte(*validPk),
+			},
+			expected: fmt.Errorf("invalid FP addr:  - empty address string is not allowed"),
 		},
 	}
 
@@ -273,6 +292,67 @@ func TestMsgEditFinalityProviderValidateBasic(t *testing.T) {
 			} else {
 				require.NoError(t, err)
 			}
+		})
+	}
+}
+
+func TestMsgSelectiveSlashingEvidence_ValidateBasic(t *testing.T) {
+	r := rand.New(rand.NewSource(10))
+	validAddr := datagen.GenRandomAddress().String()
+	validHash := datagen.GenRandomHexStr(r, 32)    // 64 chars
+	validSk := make([]byte, btcec.PrivKeyBytesLen) // 32 bytes
+
+	testCases := []struct {
+		name   string
+		msg    types.MsgSelectiveSlashingEvidence
+		expErr string
+	}{
+		{
+			name: "valid message",
+			msg: types.MsgSelectiveSlashingEvidence{
+				Signer:           validAddr,
+				StakingTxHash:    validHash,
+				RecoveredFpBtcSk: validSk,
+			},
+		},
+		{
+			name: "invalid signer address",
+			msg: types.MsgSelectiveSlashingEvidence{
+				Signer:           "not_bech32",
+				StakingTxHash:    validHash,
+				RecoveredFpBtcSk: validSk,
+			},
+			expErr: "invalid signer addr",
+		},
+		{
+			name: "invalid staking tx hash length",
+			msg: types.MsgSelectiveSlashingEvidence{
+				Signer:           validAddr,
+				StakingTxHash:    "short",
+				RecoveredFpBtcSk: validSk,
+			},
+			expErr: fmt.Sprintf("staking tx hash is not %d", chainhash.MaxHashStringSize),
+		},
+		{
+			name: "invalid BTC SK length",
+			msg: types.MsgSelectiveSlashingEvidence{
+				Signer:           validAddr,
+				StakingTxHash:    validHash,
+				RecoveredFpBtcSk: make([]byte, 16), // too short
+			},
+			expErr: fmt.Sprintf("malformed BTC SK. Expected length: %d", btcec.PrivKeyBytesLen),
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			err := tc.msg.ValidateBasic()
+			if tc.expErr == "" {
+				require.NoError(t, err)
+				return
+			}
+			require.Error(t, err)
+			require.Contains(t, err.Error(), tc.expErr)
 		})
 	}
 }